KAFKA-5682: Add `UnconsumableRecordException` interface
* Make `RecordDeserializationException` and `InoperativeRecordException` internal classes
* Add new interface that exposes topic and partition for the given exception
