+++ clients/src/main/java/org/apache/kafka/common/errors/UnconsumableRecordException.java
@@ -14,36 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.kafka.common.errors;
 
 import org.apache.kafka.common.TopicPartition;
 
 /**
- *  Any exception during deserialization in the consumer
+ * Represents an exception which was caused by a faulty record in the log.
+ * It holds information pointing to the specific record.
+ * The user is expected to seek past the offset for the given partition ({@code {@link UnconsumableRecordException#offset()} + 1}).
  */
-public class RecordDeserializationException extends SerializationException {
+public interface UnconsumableRecordException {
+    /**
+     * @return the partition of the faulty record
+     */
+    TopicPartition partition();
 
-    private static final long serialVersionUID = 1L;
-    private TopicPartition partition;
-    private long offset;
-
-    public RecordDeserializationException(TopicPartition partition, long offset, String message) {
-        super(message);
-        this.partition = partition;
-        this.offset = offset;
-    }
-
-    public TopicPartition partition() {
-        return partition;
-    }
-
-    public long offset() {
-        return offset;
-    }
-
-    /* avoid the expensive and useless stack trace for serialization exceptions */
-    @Override
-    public Throwable fillInStackTrace() {
-        return this;
-    }
+    /**
+     * @return the offset of the faulty record
+     */
+    long offset();
 }
