+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/FetcherTest.java
@@ -40,8 +40,8 @@ import org.apache.kafka.common.errors.RecordTooLargeException;
 import org.apache.kafka.common.errors.SerializationException;
 import org.apache.kafka.common.errors.TimeoutException;
 import org.apache.kafka.common.errors.TopicAuthorizationException;
-import org.apache.kafka.common.errors.RecordDeserializationException;
-import org.apache.kafka.common.errors.InoperativeRecordException;
+import org.apache.kafka.common.internals.errors.RecordDeserializationException;
+import org.apache.kafka.common.internals.errors.InoperativeRecordException;
 import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.header.internals.RecordHeader;
 import org.apache.kafka.common.metrics.KafkaMetric;
